+++ src/Data/Data/Lens.hs
@@ -14,6 +14,10 @@
 #ifdef TRUSTWORTHY
 {-# LANGUAGE Trustworthy #-}
 #endif
+-- Use -fbyte-code explicitly to ensure that -fobject-code isn't automatically
+-- implied on GHCi 8.10+ by the use of UnboxedTuples, as this breaks the
+-- doctests. See #874 for more details.
+{-# OPTIONS_GHC -fbyte-code #-}
 {-# OPTIONS_GHC -fno-full-laziness #-}
 -----------------------------------------------------------------------------
 -- |